reorganize subschema functions into subschema file
