+++ packages/delegate/src/delegateToSchema.ts
@@ -36,10 +36,10 @@ import {
   IDelegateToSchemaOptions,
   IDelegateRequestOptions,
   SubschemaConfig,
-  isSubschemaConfig,
   ExecutionParams,
   StitchingInfo,
 } from './types';
+import { isSubschemaConfig } from './Subschema';
 
 export function delegateToSchema(options: IDelegateToSchemaOptions | GraphQLSchema): any {
   if (isSchema(options)) {
