+++ packages/delegate/src/index.ts
@@ -3,10 +3,8 @@ export { createRequestFromInfo, createRequest } from './createRequest';
 export { defaultMergedResolver } from './defaultMergedResolver';
 export { createMergedResolver } from './createMergedResolver';
 export { handleResult } from './results/handleResult';
-export { getSubschema } from './subschema';
+export { Subschema, isSubschema, isSubschemaConfig, getSubschema } from './Subschema';
 
 export * from './transforms';
-
-export { Subschema } from './GraphQLSubschema';
 
 export * from './types';
