+++ packages/delegate/src/subschema.ts
@@ -1,6 +1,8 @@
 import { GraphQLSchema } from 'graphql';
 
-import { SubschemaConfig } from './types';
+import { Transform, applySchemaTransforms } from '@graphql-tools/utils';
+
+import { SubschemaConfig, MergedTypeConfig, CreateProxyingResolverFn, Subscriber, Executor } from './types';
 
 import { FIELD_SUBSCHEMA_MAP_SYMBOL, OBJECT_SUBSCHEMA_SYMBOL } from './symbols';
 
@@ -11,4 +13,34 @@ export function getSubschema(result: any, responseKey: string): GraphQLSchema | 
 
 export function setObjectSubschema(result: any, subschema: GraphQLSchema | SubschemaConfig) {
   result[OBJECT_SUBSCHEMA_SYMBOL] = subschema;
+}
+
+export function isSubschemaConfig(value: any): value is SubschemaConfig {
+  return Boolean((value as SubschemaConfig).schema);
+}
+
+export function isSubschema(value: any): value is Subschema {
+  return Boolean((value as Subschema).transformedSchema);
+}
+
+export class Subschema {
+  public schema: GraphQLSchema;
+  public rootValue?: Record<string, any>;
+  public executor?: Executor;
+  public subscriber?: Subscriber;
+  public createProxyingResolver?: CreateProxyingResolverFn;
+  public transforms: Array<Transform>;
+  public transformedSchema: GraphQLSchema;
+  public merge?: Record<string, MergedTypeConfig>;
+
+  constructor(config: SubschemaConfig) {
+    this.schema = config.schema;
+    this.executor = config.executor;
+    this.subscriber = config.subscriber;
+    this.createProxyingResolver = config.createProxyingResolver;
+    this.transforms = config.transforms ?? [];
+    this.merge = config.merge;
+
+    this.transformedSchema = applySchemaTransforms(this.schema, this.transforms);
+  }
 }
