+++ packages/delegate/src/types.ts
@@ -13,7 +13,7 @@ import {
 } from 'graphql';
 import { Operation, Transform, Request, TypeMap, ExecutionResult } from '@graphql-tools/utils';
 
-import { Subschema } from './GraphQLSubschema';
+import { Subschema } from './Subschema';
 
 export interface IDelegateToSchemaOptions<TContext = Record<string, any>, TArgs = Record<string, any>> {
   schema: GraphQLSchema | SubschemaConfig | Subschema;
@@ -124,10 +124,6 @@ export type MergedTypeResolver = (
   subschema: GraphQLSchema | SubschemaConfig,
   selectionSet: SelectionSetNode
 ) => any;
-
-export function isSubschemaConfig(value: any): value is SubschemaConfig {
-  return Boolean((value as SubschemaConfig).schema);
-}
 
 export interface StitchingInfo {
   transformedSchemas: Map<GraphQLSchema | SubschemaConfig, GraphQLSchema>;
